// src/generated/schema.ts
export const zUin = z.number().int().min(10001).max(4294967295);
// ...
export type Event = z.infer<typeof Event>;

export const FriendEntity = z.object({
  user_id: zUin.describe('用户 QQ 号'),
  nickname: z.string().describe('用户昵称'),
  sex: z.enum(['male', 'female', 'unknown']).describe('用户性别'),
  qid: z.string().describe('用户 QID'),
  remark: z.string().describe('好友备注'),
  category: z.lazy(() => FriendCategoryEntity).describe('好友分组'),
}).describe('好友实体');
export type FriendEntity = z.infer<typeof FriendEntity>;

export const FriendCategoryEntity = z.object({
  category_id: z.number().int().nonnegative().describe('好友分组 ID'),
  category_name: z.string().describe('好友分组名称'),
}).describe('好友分组实体');
export type FriendCategoryEntity = z.infer<typeof FriendCategoryEntity>;

export const GroupEntity = z.object({
  group_id: zUin.describe('群号'),
  group_name: z.string().describe('群名称'),
  member_count: z.number().int().nonnegative().describe('群成员数量'),
  max_member_count: z.number().int().nonnegative().describe('群容量'),
  remark: z.string().describe('群备注'),
  created_time: z.number().int().nonnegative().describe('群创建时间，Unix 时间戳（秒）'),
  description: z.string().describe('群简介'),
  question: z.string().describe('加群验证问题'),
  announcement: z.string().describe('群公告预览'),
}).describe('群实体');
export type GroupEntity = z.infer<typeof GroupEntity>;

export const GroupMemberEntity = z.object({
  user_id: zUin.describe('用户 QQ 号'),
  nickname: z.string().describe('用户昵称'),
  sex: z.enum(['male', 'female', 'unknown']).describe('用户性别'),
  group_id: zUin.describe('群号'),
  card: z.string().describe('成员备注'),
  title: z.string().describe('专属头衔'),
  level: z.number().int().nonnegative().describe('群等级，注意和 QQ 等级区分'),
  role: z.enum(['owner', 'admin', 'member']).describe('权限等级'),
  join_time: z.number().int().nonnegative().describe('入群时间，Unix 时间戳（秒）'),
  last_sent_time: z.number().int().nonnegative().describe('最后发言时间，Unix 时间戳（秒）'),
  shut_up_end_time: z.number().int().nonnegative().nullish().describe('禁言结束时间，Unix 时间戳（秒）'),
}).describe('群成员实体');
export type GroupMemberEntity = z.infer<typeof GroupMemberEntity>;
// ...
export const IncomingFriendMessage = z.object({
  message_scene: z.literal('friend'),
  peer_id: zUin.describe('好友 QQ 号或群号'),
  message_seq: z.number().int().nonnegative().describe('消息序列号'),
  sender_id: zUin.describe('发送者 QQ 号'),
  time: z.number().int().nonnegative().describe('消息 Unix 时间戳（秒）'),
  segments: z.array(z.lazy(() => IncomingSegment)).describe('消息段列表'),
  friend: z.lazy(() => FriendEntity).describe('好友信息'),
}).describe('好友消息');
export type IncomingFriendMessage = z.infer<typeof IncomingFriendMessage>;

export const IncomingGroupMessage = z.object({
  message_scene: z.literal('group'),
  peer_id: zUin.describe('好友 QQ 号或群号'),
  message_seq: z.number().int().nonnegative().describe('消息序列号'),
  sender_id: zUin.describe('发送者 QQ 号'),
  time: z.number().int().nonnegative().describe('消息 Unix 时间戳（秒）'),
  segments: z.array(z.lazy(() => IncomingSegment)).describe('消息段列表'),
  group: z.lazy(() => GroupEntity).describe('群信息'),
  group_member: z.lazy(() => GroupMemberEntity).describe('群成员信息'),
}).describe('群消息');
export type IncomingGroupMessage = z.infer<typeof IncomingGroupMessage>;

export const IncomingTempMessage = z.object({
  message_scene: z.literal('temp'),
  peer_id: zUin.describe('好友 QQ 号或群号'),
  message_seq: z.number().int().nonnegative().describe('消息序列号'),
  sender_id: zUin.describe('发送者 QQ 号'),
  time: z.number().int().nonnegative().describe('消息 Unix 时间戳（秒）'),
  segments: z.array(z.lazy(() => IncomingSegment)).describe('消息段列表'),
  group: z.lazy(() => GroupEntity).nullish().describe('临时会话发送者的所在的群信息'),
}).describe('临时会话消息');
export type IncomingTempMessage = z.infer<typeof IncomingTempMessage>;

export const IncomingMessage = z.discriminatedUnion('message_scene', [
  IncomingFriendMessage,
  IncomingGroupMessage,
  IncomingTempMessage,
]).describe('接收消息');
export type IncomingMessage = z.infer<typeof IncomingMessage>;
// ...
export const IncomingSegment = z.discriminatedUnion('type', [
  z.object({
    type: z.literal('text'),
    data: IncomingTextSegmentData.describe('文本消息段'),
  }).describe('文本消息段'),

  z.object({
    type: z.literal('mention'),
    data: IncomingMentionSegmentData.describe('提及消息段'),
  }).describe('提及消息段'),

  z.object({
    type: z.literal('mention_all'),
    data: IncomingMentionAllSegmentData.describe('提及全体消息段'),
  }).describe('提及全体消息段'),

  z.object({
    type: z.literal('face'),
    data: IncomingFaceSegmentData.describe('表情消息段'),
  }).describe('表情消息段'),

  z.object({
    type: z.literal('reply'),
    data: IncomingReplySegmentData.describe('回复消息段'),
  }).describe('回复消息段'),

  z.object({
    type: z.literal('image'),
    data: IncomingImageSegmentData.describe('图片消息段'),
  }).describe('图片消息段'),

  z.object({
    type: z.literal('record'),
    data: IncomingRecordSegmentData.describe('语音消息段'),
  }).describe('语音消息段'),

  z.object({
    type: z.literal('video'),
    data: IncomingVideoSegmentData.describe('视频消息段'),
  }).describe('视频消息段'),

  z.object({
    type: z.literal('file'),
    data: IncomingFileSegmentData.describe('文件消息段'),
  }).describe('文件消息段'),

  z.object({
    type: z.literal('forward'),
    data: IncomingForwardSegmentData.describe('合并转发消息段'),
  }).describe('合并转发消息段'),

  z.object({
    type: z.literal('market_face'),
    data: IncomingMarketFaceSegmentData.describe('市场表情消息段'),
  }).describe('市场表情消息段'),

  z.object({
    type: z.literal('light_app'),
    data: IncomingLightAppSegmentData.describe('小程序消息段'),
  }).describe('小程序消息段'),

  z.object({
    type: z.literal('xml'),
    data: IncomingXmlSegmentData.describe('XML 消息段'),
  }).describe('XML 消息段'),

  z.object({
    type: z.literal('markdown'),
    data: IncomingMarkdownSegmentData.describe('Markdown 消息段'),
  }).describe('Markdown 消息段'),
]).catch({
  type: 'text',
  data: { text: '[unknown]' },
}).describe('接收消息段');
export type IncomingSegment = z.infer<typeof IncomingSegment>;
// ...
export const GetLoginInfoOutput = z.object({
  uin: zUin.describe('登录 QQ 号'),
  nickname: z.string().describe('登录昵称'),
}).describe('get_login_info 响应数据');
export type GetLoginInfoOutput = z.output<typeof GetLoginInfoOutput>;
// ...
export const GetUserProfileOutput = z.object({
  nickname: z.string().describe('昵称'),
  qid: z.string().describe('QID'),
  age: z.number().int().nonnegative().describe('年龄'),
  sex: z.enum(['male', 'female', 'unknown']).describe('性别'),
  remark: z.string().describe('备注'),
  bio: z.string().describe('个性签名'),
  level: z.number().int().nonnegative().describe('QQ 等级'),
  country: z.string().describe('国家或地区'),
  city: z.string().describe('城市'),
  school: z.string().describe('学校'),
}).describe('get_user_profile 响应数据');
export type GetUserProfileOutput = z.output<typeof GetUserProfileOutput>;